confirmation messages

+++ client/pages/adminUsers.tsx
@@ -24,12 +24,15 @@ function AdminUsers() {
   }, [])
 
   const handleDelete = async (id: any) => {
-    alert("Are you sure?")
-    try{
-        await axios.delete('https://libra-tech-final-server.vercel.app/deleteUser/' + id)
-        window.location.reload()
-    }catch(err){
-        console.log(err);
+    const confirmed = window.confirm("Are you sure you want to delete this book?");
+    
+    if(confirmed){
+        try{
+            await axios.delete('https://libra-tech-final-server.vercel.app/deleteUser/' + id)
+            window.location.reload()
+        }catch(err){
+            console.log(err);
+        }
     }
   }
